+++ JS-Back-End/Cubicle/routes/auth.js
@@ -1,0 +1,13 @@
+const { Router } = require("express");
+
+const router = Router();
+
+router.get("/login", (req, res) => {
+  res.render("loginPage");
+});
+
+router.get("/register", (req, res) => {
+  res.render("registerPage");
+});
+
+module.exports = router;
